mcp-server/pyproject.toml
[project]
name = "helpdesk-mcp"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.9",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3",
    "openai>=1.30",
    "python-dotenv>=1.0",
]

[tool.uv]
package = false
